+++ src/main/java/net/bytemc/evelon/storages/Storage.java
@@ -59,9 +59,9 @@ public interface Storage {
 
     <T> boolean exists(RepositoryQuery<T> query);
 
-    <T> int count(RepositoryQuery<T> query);
+    <T> long count(RepositoryQuery<T> query);
 
-    <T> int sum(RepositoryQuery<T> query, String id);
+    <T> long sum(RepositoryQuery<T> query, String id);
 
     <T> double avg(RepositoryQuery<T> query, String id);
